ATM added
We still need to make it so that you can access the simulation from the user dashboard and we need to make all the CSS the same for all sections of the website. Also there are still bugs surrounding fund transfer that need fixing. I will be updating check deposit and account deletion soon.